backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "langfuse>=3.10.5",
    "openai-agents[litellm,viz]>=0.3.3",
    "openinference-instrumentation-openai-agents>=1.3.0",
    "playwright>=1.48.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
